[Refactor] Simplify TileLang backend directory structure
